Build Buildbarn containers using Bazel
Using Bazel to build and push the containers removes the need
to use Docker, and a privileged runner. Also, this way, the containers
produced will be built in the same way as the upstream ones.

Given that we don't need to use Docker for the build process,
we workaround the need of running `docker login` by generating the
`config.json` file manually ourselves, which is basically what it
does behind the scenes.

For now we push the images to Docker Hub, as pushing to GitLab
registry (from Bazel) is currently broken. For more details see:
<https://github.com/bazelbuild/rules_docker/issues/820>