Use buildbox-worker instead of the BuildGrid host-tools bot
The BuildGrid host-tools bot is not the recommended worker to use in a
production BuildGrid instance. The recommended approach is to use
buildbox-worker with buildbox-run-hosttools for now. There will be
further options in the future to support different sandboxing methods,
and we should add those here too when they're available.